+++ textract/parsers/__init__.py
@@ -21,9 +21,10 @@ EXTENSION_SYNONYMS = {
 # here so the default is used on both the process function and also by
 # the command line interface
 DEFAULT_ENCODING = 'utf_8'
+DEFAULT_EXTENSION = None
 
 
-def process(filename, encoding=DEFAULT_ENCODING, **kwargs):
+def process(filename, encoding=DEFAULT_ENCODING, extension=DEFAULT_EXTENSION, **kwargs):
     """This is the core function used for extracting text. It routes the
     ``filename`` to the appropriate parser and returns the extracted
     text as a byte-string encoded with ``encoding``.
@@ -36,8 +37,17 @@ def process(filename, encoding=DEFAULT_ENCODING, **kwargs):
     # get the filename extension, which is something like .docx for
     # example, and import the module dynamically using importlib. This
     # is a relative import so the name of the package is necessary
-    _, ext = os.path.splitext(filename)
+    # normally, file extension will be extracted from the file name
+    # if the file name has no extension, then the user can pass the
+    # extension as an argument
+    if extension:
+        ext = extension
+    else:
+        _, ext = os.path.splitext(filename)
     ext = ext.lower()
+    # check if the extension has the leading .
+    if not ext.startswith('.'):
+        ext = '.' + ext
 
     # check the EXTENSION_SYNONYMS dictionary
     ext = EXTENSION_SYNONYMS.get(ext, ext)